+++ src/worker/api/index.ts
@@ -49,6 +49,7 @@ import {
 	getNewLeagueLid,
 	initUILocalGames,
 	newLeagueGodModeLimits,
+	loadNames,
 } from "../util";
 import views from "../views";
 import type {
@@ -1253,6 +1254,19 @@ const getLeagueName = async () => {
 
 const getLocal = async (name: keyof Local) => {
 	return local[name];
+};
+
+const getRandomCollege = async () => {
+	// Don't use real country, since most have no colleges by default
+	const { college } = player.name("None");
+	return college;
+};
+
+const getRandomCountry = () => {
+	const playerBioInfo = local.playerBioInfo ?? loadNames();
+
+	// Equal odds of every country, otherwise it's too commonly USA - no fun!
+	return random.choice(playerBioInfo.frequencies)[0];
 };
 
 const getRandomName = async (country: string) => {
@@ -3104,6 +3118,8 @@ export default {
 	getLeagueInfo,
 	getLeagueName,
 	getLocal,
+	getRandomCollege,
+	getRandomCountry,
 	getRandomName,
 	getRandomRatings,
 	getTradingBlockOffers,